+++ packages/jivs-engine/src/Interfaces/ValidatableValueHostBase.ts
@@ -143,7 +143,7 @@ export interface IValidatableValueHostBase extends IValueHost, IGatherValueHostN
      * 
      * Use case: client-side getting server-side Jivs-generated IssuesFound,
      * so the UI can incorporate it.
-     * @param issueFound 
+     * @param issuesFound 
      * @param behavior - keep or omit an issueFound that does not have a matching validator
      * based on the errorCode.
      */