+++ packages/jivs-engine/src/Interfaces/ValidationManager.ts
@@ -158,7 +158,7 @@ export interface IValidationManager extends IValueHostsManager {
      * 
      * Use case: client-side getting server-side Jivs-generated IssuesFound,
      * so the UI can incorporate it.
-     * @param issueFound 
+     * @param issuesFound 
      * @param behavior - keep or omit an issueFound that does not have a matching validator
      * based on the errorCode.
      */