+++ packages/jivs-engine/src/Services/TextLocalizerService.ts
@@ -163,6 +163,26 @@ export class TextLocalizerService extends ServiceBase implements ITextLocalizerS
     }
 
     /**
+     * Attempts to get the localized name for a data type lookup key to be used in {DataType} token of error messages.
+     * @param dataTypeLookupKey 
+     * @returns The name or null if not available.
+     */
+    public getDataTypeLabel(cultureIdToMatch: string, dataTypeLookupKey: string): string | null
+    {
+        let text = this.localize(cultureIdToMatch, this.getDataTypeNamel10nText(dataTypeLookupKey), null);
+        if (text === null && this.fallbackService !== null)
+            return this.fallbackService.getDataTypeLabel(cultureIdToMatch, dataTypeLookupKey);  
+        if (text === null && dataTypeLookupKey)
+            return dataTypeLookupKey;
+        return text;
+    }    
+
+    protected getDataTypeNamel10nText(dataTypeLookupKey: string): string
+    {
+        return 'DTLK-' + dataTypeLookupKey;
+    }    
+
+    /**
      * Registers a lookup key with the culture specific text.
      * Replaces an already registered entry with the same l10nKey.
      * @param l10nKey - Localization key, which is the text that identifies which word,
@@ -201,6 +221,20 @@ export class TextLocalizerService extends ServiceBase implements ITextLocalizerS
     {
         this.register(this.getSummaryMessagel10nText(errorCode, dataTypeLookupKey), cultureToText);
     }    
+
+    /**
+     * Utility to add text representation of a data type associating it with its 
+     * dataTypeLookupKey. The text is used with the {DataType} token in error messages.
+     * The localization key (l10ntext) will use this pattern:
+     * 'DTLK-' + DataTypeLookupKey
+     * @param dataTypeLookupKey
+     * @param cultureToText 
+     */
+    public registerDataTypeLabel(dataTypeLookupKey: string, cultureToText: CultureToText): void
+    {
+        this.register(this.getDataTypeNamel10nText(dataTypeLookupKey), cultureToText);
+    }
+
 
     /**
      * Data is stored here, where the key is the l10nKey and the value