+++ packages/jivs-engine/src/ValueHosts/ValidatableValueHostBase.ts
@@ -462,7 +462,7 @@ export abstract class ValidatableValueHostBase<TConfig extends ValidatableValueH
      * 
      * Use case: client-side getting server-side Jivs-generated IssuesFound,
      * so the UI can incorporate it.
-     * @param issueFound 
+     * @param issuesFound 
      * @param behavior - keep or omit an issueFound that does not have a matching validator
      * based on the errorCode. Defaults to Keep.
      */